Clean renamincs for \services
